requirements.txt
requests
pandas
rich
schedule
sqlite3
smtplib
twilio
python-telegram-bot